refactor(#606): organize controllers into folders
