+++ app/Http/Controllers/Admin/CommentsController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\BookingRequest;
 use App\Models\Comment;
 use Illuminate\Http\Request;
