+++ app/Http/Controllers/Admin/RestrictionsController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\Room;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Http\RedirectResponse;
