+++ app/Http/Controllers/Admin/SettingsController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\AcademicDate;
 use App\Models\Settings;
 use Illuminate\Http\RedirectResponse;
