+++ app/Http/Controllers/System/DashboardController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\System;
 
+use App\Http\Controllers\Controller;
 use App\Models\Room;
 use App\Models\BookingRequest;
 use App\Models\User;
