+++ app/Http/Controllers/System/RoomController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\System;
 
+use App\Http\Controllers\Controller;
 use App\Http\Resources\RoomCollection;
 use App\Http\Resources\RoomResource;
 use App\Models\Availability;
